+++ src/services/authentication-service/open-id/open-id-url-utility.ts
