+++ src/services/authentication-service/web.oidc.authentication.service.ts
@@ -8,7 +8,7 @@ import {User} from 'oidc-client';
 
 import {AuthenticationStateEvent, IAuthenticationService, ILoginResult, IUserIdentity, NotificationType} from '../../contracts/index';
 import {oidcConfig} from '../../open-id-connect-configuration';
-import {NotificationService} from './../notification/notification.service';
+import {NotificationService} from './../../services/notification-service/notification.service';
 
 const UNAUTHORIZED_STATUS_CODE: number = 401;
 const IDENTITY_SERVER_AVAILABLE_SUCCESS_STATUS_CODE: number = 200;
