+++ src/services/diagram-creation-service/DiagramCreationService.ts
