+++ src/services/diagram-validation-service/DiagramValidationService.ts
