+++ src/services/diagram-validation-service/DiagramValidator.ts
