+++ src/services/diagram-validation-service/index.ts
