+++ src/services/solution-explorer-services/DiagramTrashFolderService.ts
