+++ src/services/solution-explorer-services/SingleDiagramsSolutionExplorerService.ts
